second

+++ src/main/java/tn/esprit/etudiantex/entities/Bloc.java
@@ -18,7 +18,7 @@ public class Bloc {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idBloc")
-    private long idBloc;
+    private Integer idBloc;
     @Column(name = "nomBloc")
     private String nomBloc;
     @Column(name = "capaciteBloc")
